requirements.txt
aws-cdk-lib==1.187.0
constructs>=10.0.0,<11.0.0
